requirements.txt
# Core Dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# VirtualHome Dependencies
# IMPORTANT: Use numpy<2.0 for compatibility with opencv-python
certifi==2024.7.4
chardet==3.0.4
idna==3.7
matplotlib>=3.4.2
networkx==2.3
numpy<2.0
opencv-python==4.10.0.84
pillow>=8.3.1
plotly==3.10.0
requests==2.32.2
ipdb==0.13.9
termcolor==1.1.0
tqdm>=4.67.1
urllib3>=1.24.3

# Python Version: 3.10+ (tested with 3.10 and 3.13)

# Video Generation
# Requires FFmpeg installed separately:
# macOS: brew install ffmpeg
# Ubuntu: apt-get install ffmpeg
# Windows: Download from https://ffmpeg.org/download.html